# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  needsLogin: boolean;
  /** @nullable */
  uptime: number | null;
  /** @nullable */
  screenshot: string | null;
  /** @nullable */
  lastActivity: string | null;
  sessionActive: boolean;
  earning: boolean;
  /** @nullable */
  lastVerification: string | null;
  verificationCount: number;
}

export interface BotActionResult {
  success: boolean;
  message: string;
}

export interface ClickRequest {
  x: number;
  y: number;
}

export interface TypeRequest {
  text: string;
}

export interface ScreenshotResponse {
  /** @nullable */
  screenshot: string | null;
  timestamp: string;
}