# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/click:
    post:
      operationId: botClick
      tags: [bot]
      summary: Click at coordinates in the browser
      description: Simulate a mouse click at given x,y coordinates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClickRequest"
      responses:
        "200":
          description: Click performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotActionResult"
  /bot/type:
    post:
      operationId: botType
      tags: [bot]
      summary: Type text in the browser
      description: Types text at the current cursor position in the browser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TypeRequest"
      responses:
        "200":
          description: Text typed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotActionResult"
...
components:
  schemas:
...
    BotActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ClickRequest:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
      required:
        - x
        - y
    TypeRequest:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...